d_a_tbox: Match leftover actionOpenWait (demo partial)

+++ include/d/actor/d_a_tbox.h
@@ -34,7 +34,7 @@ public:
 
     int getTboxNo() { return fopAcM_GetParam(this) >> 0x07 & 0x1F; }
     int getSwNo() { return fopAcM_GetParam(this) >> 0x0C & 0xFF; }
-    u8 getItemNo() { return home.angle.z >> 8 & 0xFF; }
+    int getItemNo() { return home.angle.z >> 8 & 0xFF; }
     void flagOn(u16 flag) { mFlags |= flag; }
     void flagOff(u16 flag) { mFlags &= ~flag; }
     void flagClr() { mFlags = 0; }
