+++ src/d/actor/d_a_tbox.cpp
@@ -957,7 +957,7 @@ BOOL daTbox_c::actionOpenWait() {
     if (eventInfo.checkCommandDoor()) {
         dComIfGp_event_onEventFlag(0x04);
 
-        u8 itemNo = getItemNo();
+        int itemNo = getItemNo();
         fpc_ProcID itemPID = fopAcM_createItemForTrBoxDemo(&current.pos, itemNo);
 
         if (itemPID != fpcM_ERROR_PROCESS_ID_e) {
